+++ fjs/ci/module.f.ts
@@ -7,17 +7,19 @@ import { mapStep, step, type Effect } from '../effects/module.f.ts'
 import { access, writeUtf8File, type NodeOp } from '../effects/node/module.f.ts'
 import { functionalscript, images } from './config/module.f.mjs'
 import {
-    type Architecture,
-    type GitHubAction,
-    type Job,
-    type Jobs,
-    type MetaStep,
-    type Os,
     architecture,
     os,
     toSteps,
     ubuntuArm
-} from './common/module.f.ts'
+} from './common/module.f.mjs'
+import type {
+    Architecture,
+    GitHubAction,
+    Job,
+    Jobs,
+    MetaStep,
+    Os,
+} from './common/types.ts'
 import { rustPlatformSteps, rustWasmSteps } from './rust/module.f.ts'
 import { nodeMainSteps, nodeNixJobs, nodeNixVersionSteps, nodeVersionJobs } from './node/module.f.ts'
 import { nixFlakes, nixInstall, type NixJob } from './nix/module.f.ts'