+++ fjs/ci/rust/module.f.ts
@@ -7,7 +7,8 @@
  * @module
  */
 import { wasmer, wasmtime } from '../config/module.f.mjs'
-import { type Architecture, type MetaStep, type Os, install, test, uses } from '../common/module.f.ts'
+import { install, test, uses } from '../common/module.f.mjs'
+import type { Architecture, MetaStep, Os } from '../common/types.ts'
 
 const cargoCommand = (tool: 'clippy' | 'test', target?: string, config?: string): string => {
     const to = target ? ` --target ${target}` : ''